nw_sample/grpc_other_lang/server/requirements.txt
grpcio==1.24.1
grpcio-tools==1.24.1
protobuf==3.9.1